phase-3/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
SQLAlchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib
bcrypt==4.2.0
python-multipart
python-dotenv
pydantic
openai
mcp
fastmcp
google-generativeai

# Pin httpx for Starlette TestClient compatibility (Starlette 0.27 expects httpx<0.28)
httpx<0.28
